fix: enforce frozen image inputs at the canonical pack boundary
